+++ lib/src/utils/helpers.dart
@@ -1,28 +1,23 @@
 import 'dart:convert';
-import 'dart:developer';
-import 'dart:io';
 
 import 'package:dio/dio.dart';
 import 'package:file_saver/src/models/link_details.dart';
+import 'package:file_saver/src/utils/file_ops_stub.dart'
+    if (dart.library.io) 'package:file_saver/src/utils/file_ops_io.dart'
+    as file_ops;
 import 'package:flutter/foundation.dart';
-import 'package:path_provider/path_provider.dart' as path_provider;
-import 'package:path_provider_linux/path_provider_linux.dart'
-    as path_provider_linux;
-import 'package:path_provider_windows/path_provider_windows.dart'
-    as path_provder_windows;
 
 ///Helper Class for serveral utility methods
 ///
 class Helpers {
   ///This method provides [Uint8List] from [File]
-  static Future<Uint8List> _getBytesFromFile(File file) async {
-    return await file.readAsBytes();
+  static Future<Uint8List> _getBytesFromFile(Object file) async {
+    return file_ops.readFileBytes(file);
   }
 
   ///This method provides [Uint8List] from file path
   static Future<Uint8List> _getBytesFromPath(String path) async {
-    File file = File(path);
-    return await file.readAsBytes();
+    return file_ops.readPathBytes(path);
   }
 
   ///This method provides [Uint8List] from link
@@ -35,7 +30,8 @@ class Helpers {
     Dio? dioClient,
     Uint8List Function(dynamic data)? transformDioResponse,
   }) async {
-    final dio = dioClient ??
+    final dio =
+        dioClient ??
         Dio(
           BaseOptions(
             headers: link.headers,
@@ -67,35 +63,11 @@ class Helpers {
 
   ///This method provides default downloads directory for saving the file for Android, iOS, Linux, Windows, macOS
   static Future<String?> getDirectory() async {
-    String? path;
-    try {
-      if (Platform.isIOS || Platform.isAndroid) {
-        path = (await path_provider.getApplicationDocumentsDirectory()).path;
-      } else if (Platform.isMacOS) {
-        path = (await path_provider.getDownloadsDirectory())?.path;
-      } else if (Platform.isWindows) {
-        path_provder_windows.PathProviderWindows pathWindows =
-            path_provder_windows.PathProviderWindows();
-        path = await pathWindows.getDownloadsPath();
-      } else if (Platform.isLinux) {
-        path_provider_linux.PathProviderLinux pathLinux =
-            path_provider_linux.PathProviderLinux();
-        path = await pathLinux.getDownloadsPath();
-      }
-    } on Exception catch (e) {
-      log('Something wemt worng while getting directories');
-      log(e.toString());
-      rethrow;
-    }
-    return path;
+    return file_ops.getDirectory();
   }
 
   static String getFilePathSlash() {
-    if (Platform.isWindows) {
-      return '\\';
-    } else {
-      return '/';
-    }
+    return '/';
   }
 
   ///This method is used to format the extension as per the requirement
@@ -114,12 +86,14 @@ class Helpers {
   static Future<Uint8List> getBytes({
     String? filePath,
     LinkDetails? link,
-    File? file,
+    Object? file,
     Dio? dioClient,
     Uint8List Function(dynamic data)? transformDioResponse,
   }) async {
-    assert(filePath != null || link != null || file != null,
-        'Either filePath or link or file must be provided');
+    assert(
+      filePath != null || link != null || file != null,
+      'Either filePath or link or file must be provided',
+    );
     if (filePath != null) {
       return _getBytesFromPath(filePath);
     } else {